Pushover notifications + state
My updated version at home uses `state = x` instead of `fan_on = True / False` which is more efficient. I have also provided functions to trigger push notifications, using Pushover, to alert you when the fan is turned on (i.e. because it's too hot) and when it's cool enough that the fan has turned off. This may be quite useful so that you aren't worrying about knowing if the fan is on or not. You just need to sign up for Pushover (£3.99 one off fee usually) and have the app on your device.